Add files via upload
In this notebook,  the neighborhoods in the city of Toronto are explored, segmented, and clustered. 
Following libraries were used: 
PANDAS, NUMPY, MATPLOTLIB, SCIKIT LEARN, FOLIUM, GEOPY

Postal codes, Boroughs, and Neighborhood data were scraped from the table given in the following Wikipedia page, https://en.wikipedia.org/wiki/List_of_postal_codes_of_Canada:_M. 
data was then transformed into a pandas dataframe. 
The data is then cleaned and prepared for further analysis. 
From geospatial_cordinate.csv file latitude and longitude columns for respective postal codes were collected and added to the pandas dataframe so that it contains the following columns 
Postal Codes, 
Borough, 
Neighborhood,
Latitude and
 Longitude

For the Toronto neighborhood data, a Wikipedia page exists that has all the information we need to explore and cluster the neighborhoods in Toronto. You will be required to scrape the Wikipedia page and wrangle the data, clean it, and then read it into a pandas dataframe so that it is in a structured format like the New York dataset.

Once the data is in a structured format, you can replicate the analysis that we did to the New York City dataset to explore and cluster the neighborhoods in the city of Toronto.

Latitude and longitude  of Scarborough also found using  Nominatim module from geopy geocoder
A map of scarborough was plotted using folium. 

At this point, we are ready to fetch information about all the venues in the neighborhood areas in Toronto. For that reason, Forsqure Credentials and Versions are defined. 
Then a function namely 'getNearbyVenues' is written that takes four parameters, uses the requests model and stores the result in json format. 
The new dataframe, scarborough_venues has 7 columns. They are:
    Neighborhood
    Neighborhood Latitude
    Neighborhood Longitude
    Venue
    Venue Lattitude 
    Venue Longitude
    Venue Category
The data shows that  There are 273 uniques categories of venue in the Neighborhoods.
K-means clustering was used to find the cluster of venues.